TimedText and GameInfo now have their own files.
