+++ src/Miscellaneous.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "TimedText.h"
 #include "raylib.h"
 #include <functional>
 #include <vector>
@@ -16,28 +17,6 @@ enum class RotationDirection
 {
     CCW,
     CW
-};
-
-struct GameInfo {
-    int score = 0;
-    int faceHits = 0;
-    std::vector<float> possibleSpeeds = {
-        1.0f, 1.5f, 2.0f, 2.5f, 3.0f, 5.0f, 6.0f, 7.5, 9.0f
-    };
-    int anger = 0;
-    int rotationCountdown = 20;
-};
-
-class TimedText
-{
-public:
-    const std::function<void()> drawFunction;
-    float lastCheckTime;
-    float duration;
-    const char* text;
-
-    void Draw();
-    TimedText(const std::function<void()> drawFunction);
 };
 
 namespace Utilities
